requirements.txt
py-cord[voice]==2.6.0
py-cord[speed]==2.6.0
python-dotenv==1.0.1
yt_dlp==2024.7.9
python-dotenv==1.0.1
Nueue==1.0.0